+++ test/api/api_test.cpp
@@ -696,7 +696,7 @@ TEST_F(APITest, module3) {
 TEST_F(APITest, module4) {
     auto fileName = this->createTempFile("target.ds", "source hoghreua");
     DSError e;
-    int r = DSState_loadModule(this->state, fileName.c_str(), DS_MOD_FULLPATH | DS_MOD_IGNORE_ENOENT, &e);
+    int r = DSState_loadModule(this->state, fileName.c_str(), DS_MOD_IGNORE_ENOENT, &e);
     ASSERT_EQ(1, r);
     ASSERT_EQ(DS_ERROR_KIND_TYPE_ERROR, e.kind);
     ASSERT_STREQ("NotFoundMod", e.name);
@@ -706,7 +706,7 @@ TEST_F(APITest, module4) {
     // check error message
     auto ret = invoke([&]{
         return DSState_loadModule(this->state, fileName.c_str(),
-                DS_MOD_FULLPATH | DS_MOD_IGNORE_ENOENT, nullptr);
+                DS_MOD_IGNORE_ENOENT, nullptr);
     });
     ASSERT_NO_FATAL_FAILURE(this->expectRegex(ret, 1, WaitStatus::EXITED, "",
                                          "^.+/target.ds:1: \\[semantic error\\] module not found: `hoghreua'.+$"));
@@ -714,7 +714,7 @@ TEST_F(APITest, module4) {
 
 TEST_F(APITest, module5) {
     DSError e;
-    int r = DSState_loadModule(this->state, "hfeurhfiurhefuie", DS_MOD_FULLPATH , &e);
+    int r = DSState_loadModule(this->state, "hfeurhfiurhefuie", 0, &e);
     ASSERT_EQ(1, r);
     ASSERT_EQ(DS_ERROR_KIND_TYPE_ERROR, e.kind);
     ASSERT_STREQ("NotFoundMod", e.name);
@@ -723,8 +723,7 @@ TEST_F(APITest, module5) {
 
     // check error message
     auto ret = invoke([&]{
-        return DSState_loadModule(this->state, "freijjfeir",
-                                  DS_MOD_FULLPATH, nullptr);
+        return DSState_loadModule(this->state, "freijjfeir", 0, nullptr);
     });
     ASSERT_NO_FATAL_FAILURE(this->expect(ret, 1, WaitStatus::EXITED, "",
                     "ydsh: [semantic error] module not found: `freijjfeir'"));
@@ -739,7 +738,7 @@ TEST_F(APITest, module6) {
         \ \
 )", "echo moduel!!; exit 56");
     auto ret = invoke([&]{
-        return DSState_loadModule(this->state, fileName.c_str(), DS_MOD_FULLPATH, nullptr);
+        return DSState_loadModule(this->state, fileName.c_str(), 0, nullptr);
     });
     ASSERT_NO_FATAL_FAILURE(this->expect(ret, 56, WaitStatus::EXITED, "moduel!!"));
 }