+++ test/interactive/interactive_test.cpp
@@ -533,7 +533,8 @@ TEST_F(InteractiveTest, rc3) {
 TEST_F(InteractiveTest, rc4) {
     this->invoke("--quiet", "--rcfile", ".");
     ASSERT_NO_FATAL_FAILURE(this->waitAndExpect(1, WaitStatus::EXITED, "",
-            "ydsh: [semantic error] cannot read module: `.', by `Is a directory'\n"));
+            format("ydsh: [semantic error] cannot read module: `%s', by `Is a directory'\n",
+                   INTERACTIVE_TEST_WORK_DIR)));
 }
 
 TEST_F(InteractiveTest, rc5) {